Factor out `start_reconnect_task`.
